backend/breathesync-backend/requirements.txt
fastapi
uvicorn[standard]
librosa
numpy
python-multipart
scikit-learn
scipy
httpx
python-dotenv
soundfile
